+++ spec/smarter_csv/parse/column_separator_spec.rb
@@ -12,15 +12,24 @@
 #   code paths are behaving identically.
 # ------------------------------------------------------------------------------------------
 
+class Klass
+  include SmarterCSV::Parser
+  def has_acceleration
+    true
+  end
+end
+
 [true, false].each do |bool|
   describe "fulfills RFC-4180 and more with#{bool ? ' C-' : 'out '}acceleration" do
+    let(:instance) { Klass.new }
+
     describe 'parse with col_sep' do
       let(:options) { {quote_char: '"', acceleration: bool} }
 
       it 'parses with comma' do
         line = "a,b,,d"
         options.merge!({col_sep: ","})
-        array, array_size = SmarterCSV.send(:parse, line, options)
+        array, array_size = instance.send(:parse, line, options)
         expect(array).to eq ['a', 'b', '', 'd']
         expect(array_size).to eq 4
       end
@@ -28,7 +37,7 @@
       it 'parses trailing commas' do
         line = "a,b,c,,"
         options.merge!({col_sep: ","})
-        array, array_size = SmarterCSV.send(:parse, line, options)
+        array, array_size = instance.send(:parse, line, options)
         expect(array).to eq ['a', 'b', 'c', '', '']
         expect(array_size).to eq 5
       end
@@ -36,7 +45,7 @@
       it 'parses with space' do
         line = "a b  d"
         options.merge!({col_sep: " "})
-        array, array_size = SmarterCSV.send(:parse, line, options)
+        array, array_size = instance.send(:parse, line, options)
         expect(array).to eq ['a', 'b', '', 'd']
         expect(array_size).to eq 4
       end
@@ -44,7 +53,7 @@
       it 'parses with tab' do
         line = "a\tb\t\td"
         options.merge!({col_sep: "\t"})
-        array, array_size = SmarterCSV.send(:parse, line, options)
+        array, array_size = instance.send(:parse, line, options)
         expect(array).to eq ['a', 'b', '', 'd']
         expect(array_size).to eq 4
       end
@@ -52,7 +61,7 @@
       it 'parses with multiple space separator' do
         line = "a b    d"
         options.merge!({col_sep: "  "})
-        array, array_size = SmarterCSV.send(:parse, line, options)
+        array, array_size = instance.send(:parse, line, options)
         expect(array).to eq ['a b', '', 'd']
         expect(array_size).to eq 3
       end
@@ -60,7 +69,7 @@
       it 'parses with multiple char separator' do
         line = '<=><=>A<=>B<=>C'
         options.merge!({col_sep: "<=>"})
-        array, array_size = SmarterCSV.send(:parse, line, options)
+        array, array_size = instance.send(:parse, line, options)
         expect(array).to eq ["", "", "A", "B", "C"]
         expect(array_size).to eq 5
       end
@@ -68,7 +77,7 @@
       it 'parses trailing multiple char separator' do
         line = '<=><=>A<=>B<=>C<=><=>'
         options.merge!({col_sep: "<=>"})
-        array, array_size = SmarterCSV.send(:parse, line, options)
+        array, array_size = instance.send(:parse, line, options)
         expect(array).to eq ["", "", "A", "B", "C", "", ""]
         expect(array_size).to eq 7
       end