+++ spec/smarter_csv/parse/old_csv_library_spec.rb
@@ -12,8 +12,17 @@
 #   code paths are behaving identically.
 # ------------------------------------------------------------------------------------------
 
+class Klass
+  include SmarterCSV::Parser
+  def has_acceleration
+    true
+  end
+end
+
 [true, false].each do |bool|
   describe "fulfills RFC-4180 and more with#{bool ? ' C-' : 'out '}acceleration" do
+    let(:instance) { Klass.new }
+
     describe 'old CSV library parsing tests' do
       let(:options) { {quote_char: '"', col_sep: ",", acceleration: bool} }
 
@@ -45,7 +54,7 @@
        ["foo,\"foo,bar\",baz", ["foo", "foo,bar", "baz"]],
        [";,;", [";", ";"]]].each do |line, result|
         it "parses #{line} as #{result.inspect}" do
-          array, _array_size = SmarterCSV.send(:parse, line, options)
+          array, _array_size = instance.send(:parse, line, options)
           expect(array).to eq result
         end
       end
@@ -67,26 +76,26 @@
        ["foo,\"\r\n\n\",baz", ["foo", "\r\n\n", "baz"]],
        ["foo,\"foo,bar\",baz", ["foo", "foo,bar", "baz"]]].each do |line, result|
         it "parses #{line} as #{result.inspect}" do
-          array, _array_size = SmarterCSV.send(:parse, line, options)
+          array, _array_size = instance.send(:parse, line, options)
           expect(array).to eq result
         end
       end
 
       it 'quoted test' do
         line = '"This",is,2"","3""",test'
-        array, _array_size = SmarterCSV.send(:parse, line, options)
+        array, _array_size = instance.send(:parse, line, options)
         expect(array).to eq ["This", "is", '2"', '3"', "test"]
       end
 
       it 'mixed quotes' do
         line = %{Ten Thousand,10000, 2710 ,,"10,000","It's ""10 Grand"", baby",10K}
-        array, _array_size = SmarterCSV.send(:parse, line, options)
+        array, _array_size = instance.send(:parse, line, options)
         expect(array).to eq ["Ten Thousand", "10000", " 2710 ", "", "10,000", "It's \"10 Grand\", baby", "10K"]
       end
 
       it 'single quotes in fields' do
         line = 'Indoor Chrome,49.2"" L x 49.2"" W x 20.5"" H,Chrome,"Crystal,Metal,Wood",23.12'
-        array, _array_size = SmarterCSV.send(:parse, line, options)
+        array, _array_size = instance.send(:parse, line, options)
         expect(array).to eq ['Indoor Chrome', '49.2" L x 49.2" W x 20.5" H', 'Chrome', 'Crystal,Metal,Wood', '23.12']
       end
     end