+++ spec/smarter_csv/parse/rfc4180_and_more_spec.rb
@@ -12,20 +12,28 @@
 #   code paths are behaving identically.
 # ------------------------------------------------------------------------------------------
 
+class Klass
+  include SmarterCSV::Parser
+  def has_acceleration
+    true
+  end
+end
+
 [true, false].each do |bool|
   describe "fulfills RFC-4180 and more with#{bool ? ' C-' : 'out '}acceleration" do
     let(:options) { {col_sep: ',', row_sep: $INPUT_RECORD_SEPARATOR, quote_char: '"', acceleration: bool } }
+    let(:instance) { Klass.new }
 
     context 'parses simple CSV' do
       context 'RFC-4180' do
         it 'separating on col_sep' do
           line = 'aaa,bbb,ccc'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [%w[aaa bbb ccc], 3]
+          expect(instance.send(:parse, line, options)).to eq [%w[aaa bbb ccc], 3]
         end
 
         it 'preserves whitespace' do
           line = ' aaa , bbb , ccc '
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             [' aaa ', ' bbb ', ' ccc '], 3
           ]
         end
@@ -34,42 +42,42 @@
       context 'extending RFC-4180' do
         it 'with extra col_sep' do
           line = 'aaa,bbb,ccc,'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             ['aaa', 'bbb', 'ccc', ''], 4
           ]
         end
 
         it 'with extra col_sep with given header_size' do
           line = 'aaa,bbb,ccc,'
-          expect(SmarterCSV.send(:parse, line, options, 3)).to eq [
+          expect(instance.send(:parse, line, options, 3)).to eq [
             %w[aaa bbb ccc], 3
           ]
         end
 
         it 'with multiple extra col_sep' do
           line = 'aaa,bbb,ccc,,,'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             ['aaa', 'bbb', 'ccc', '', '', ''], 6
           ]
         end
 
         it 'with multiple extra col_sep' do
           line = 'aaa,bbb,ccc,,,'
-          expect(SmarterCSV.send(:parse, line, options, 3)).to eq [
+          expect(instance.send(:parse, line, options, 3)).to eq [
             %w[aaa bbb ccc], 3
           ]
         end
 
         it 'with multiple complex col_sep' do
           line = 'aaa<=>bbb<=>ccc<=><=><=>'
-          expect(SmarterCSV.send(:parse, line, options.merge({col_sep: '<=>'}))).to eq [
+          expect(instance.send(:parse, line, options.merge({col_sep: '<=>'}))).to eq [
             ['aaa', 'bbb', 'ccc', '', '', ''], 6
           ]
         end
 
         it 'with multiple complex col_sep with given header_size' do
           line = 'aaa<=>bbb<=>ccc<=><=><=>'
-          expect(SmarterCSV.send(:parse, line, options.merge({col_sep: '<=>'}), 3)).to eq [
+          expect(instance.send(:parse, line, options.merge({col_sep: '<=>'}), 3)).to eq [
             %w[aaa bbb ccc], 3
           ]
         end
@@ -80,47 +88,47 @@
       context 'RFC-4180' do
         it 'separating on col_sep' do
           line = '"aaa","bbb","ccc"'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [%w[aaa bbb ccc], 3]
+          expect(instance.send(:parse, line, options)).to eq [%w[aaa bbb ccc], 3]
         end
 
         it 'parses corner case correctly' do
           line = '"Board 4""","$17.40","10000003427"'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             ['Board 4"', '$17.40', '10000003427'], 3
           ]
         end
 
         it 'quoted parts can contain spaces' do
           line = '" aaa1 aaa2 "," bbb1 bbb2 "," ccc1 ccc2 "'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             [' aaa1 aaa2 ', ' bbb1 bbb2 ', ' ccc1 ccc2 '], 3
           ]
         end
 
         it 'quoted parts can contain row_sep' do
           line = '"aaa1, aaa2","bbb1, bbb2","ccc1, ccc2"'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             ['aaa1, aaa2', 'bbb1, bbb2', 'ccc1, ccc2'], 3
           ]
         end
 
         it 'quoted parts can contain row_sep' do
           line = '"aaa1, ""aaa2"", aaa3","""bbb1"", bbb2","ccc1, ""ccc2"""'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             ['aaa1, "aaa2", aaa3', '"bbb1", bbb2', 'ccc1, "ccc2"'], 3
           ]
         end
 
         it 'some fields are quoted' do
           line = '1,"board 4""",12.95'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             ['1', 'board 4"', '12.95'], 3
           ]
         end
 
         it 'separating on col_sep' do
           line = '"some","thing","""completely"" different"'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             ['some', 'thing', '"completely" different'], 3
           ]
         end
@@ -129,40 +137,40 @@
       context 'extending RFC-4180' do
         it 'with extra col_sep, without given header_size' do
           line = '"aaa","bbb","ccc",'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             ['aaa', 'bbb', 'ccc', ''], 4
           ]
         end
 
         it 'with extra col_sep, with given header_size' do
           line = '"aaa","bbb","ccc",'
-          expect(SmarterCSV.send(:parse, line, options, 3)).to eq [%w[aaa bbb ccc], 3]
+          expect(instance.send(:parse, line, options, 3)).to eq [%w[aaa bbb ccc], 3]
         end
 
         it 'with multiple extra col_sep, without given header_size' do
           line = '"aaa","bbb","ccc",,,'
-          expect(SmarterCSV.send(:parse, line, options)).to eq [
+          expect(instance.send(:parse, line, options)).to eq [
             ['aaa', 'bbb', 'ccc', '', '', ''], 6
           ]
         end
 
         it 'with multiple extra col_sep, with given header_size' do
           line = '"aaa","bbb","ccc",,,'
-          expect(SmarterCSV.send(:parse, line, options, 3)).to eq [
+          expect(instance.send(:parse, line, options, 3)).to eq [
             %w[aaa bbb ccc], 3
           ]
         end
 
         it 'with multiple complex extra col_sep, without given header_size' do
           line = '"aaa"<=>"bbb"<=>"ccc"<=><=><=>'
-          expect(SmarterCSV.send(:parse, line, options.merge({col_sep: '<=>'}))).to eq [
+          expect(instance.send(:parse, line, options.merge({col_sep: '<=>'}))).to eq [
             ['aaa', 'bbb', 'ccc', '', '', ''], 6
           ]
         end
 
         it 'with multiple complex extra col_sep, with given header_size' do
           line = '"aaa"<=>"bbb"<=>"ccc"<=><=><=>'
-          expect(SmarterCSV.send(:parse, line, options.merge({col_sep: '<=>'}), 3)).to eq [
+          expect(instance.send(:parse, line, options.merge({col_sep: '<=>'}), 3)).to eq [
             %w[aaa bbb ccc], 3
           ]
         end
@@ -173,7 +181,7 @@
     context 'liberal_parsing' do
       it 'parses corner case correctly' do
         line = 'is,this "three, or four",fields'
-        expect(SmarterCSV.send(:parse, line, options)).to eq [
+        expect(instance.send(:parse, line, options)).to eq [
           ['is', 'this "three, or four"', 'fields'], 3
         ]
       end